+++ src/core/tusClientReducer.ts
@@ -1,5 +1,6 @@
 import { Reducer } from 'react';
 import { Upload } from 'tus-js-client';
+import { TusClientActions } from './tucClientActions';
 
 export type UploadState = {
   upload: Upload | undefined;
@@ -12,50 +13,6 @@ export type TusClientState = {
     [uploadKey: string]: UploadState | undefined;
   };
 };
-
-export type TusClientActions = ReturnType<
-  | typeof insertUploadInstance
-  | typeof removeUploadInstance
-  | typeof successUpload
-  | typeof errorUpload
->;
-
-export const insertUploadInstance = (uploadKey: string, upload: Upload) =>
-  ({
-    type: 'INSERT_UPLOAD_INSTANCE',
-    payload: {
-      uploadKey,
-      uploadState: {
-        upload,
-        isSuccess: false,
-      },
-    },
-  } as const);
-
-export const successUpload = (uploadKey: string) =>
-  ({
-    type: 'SUCCESS_UPLOAD',
-    payload: {
-      uploadKey,
-    },
-  } as const);
-
-export const errorUpload = (uploadKey: string, error?: Error) =>
-  ({
-    type: 'ERROR_UPLOAD',
-    payload: {
-      uploadKey,
-      error,
-    },
-  } as const);
-
-export const removeUploadInstance = (uploadKey: string) =>
-  ({
-    type: 'REMOVE_UPLOAD_INSTANCE',
-    payload: {
-      uploadKey,
-    },
-  } as const);
 
 export const tusClientReducer: Reducer<TusClientState, TusClientActions> = (
   state,
